+++ Tapjoy/CHANGELOG.md
@@ -1,6 +1,11 @@
 ## Changelog
+  * 12.0.0.0
+    * This version of the adapters has been certified with Tapjoy 12.0.0.
+    * Add `TapjoyAdvancedBidder` for publishers using Advanced Bidding.
+
   * 11.12.2.1
     * Update adapters to be compatible with MoPub iOS SDK framework
+    
   * 11.12.2.0
     * This version of the adapters has been certified with Tapjoy 11.12.2.
     * General Data Protection Regulation (GDPR) update to support a way for publishers to determine GDPR applicability and to obtain/manage consent from users in European Economic Area, the United Kingdom, or Switzerland to serve personalize ads. Only applicable when integrated with MoPub version 5.0.0 and above.